refactor: ♻️ move analysis logic away from react component
